+++ api/src/main/java/com/clinics_schedules/clinic_api/repository/BuildingRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.clinics_schedules.clinic_api.entity.Building;
 import com.clinics_schedules.clinic_api.interfaces.FindableByName;
 
-public interface BuildingRepository extends JpaRepository<Building ,Long> ,FindableByName<Building> {
+public interface BuildingRepository extends JpaRepository<Building ,Integer> ,FindableByName<Building> {
      
 }
