+++ api/src/main/java/com/clinics_schedules/clinic_api/repository/EmployeeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.clinics_schedules.clinic_api.entity.Employee;
 import com.clinics_schedules.clinic_api.interfaces.FindableByName;
 
-public interface EmployeeRepository extends JpaRepository<Employee ,  Long> ,FindableByName<Employee>{
+public interface EmployeeRepository extends JpaRepository<Employee ,  Integer> ,FindableByName<Employee>{
     
 }
